requirements.txt
slack-sdk==3.27.1
python-dotenv==1.0.0
watchdog==4.0.0
requests==2.31.0
aiohttp==3.12.15
flask==3.0.0
flask-cors==4.0.0
langchain==0.3.27
langchain-openai==0.3.28
langchain-community==0.3.27
cffi>=1.15.0
argcomplete>=1.9.4
cryptography>=3.4.8
pynacl>=1.5.0
# Additional security dependencies
PyJWT==2.8.0
pyopenssl==23.3.0
# Production monitoring dependencies
psutil>=5.9.0